Add a 'hostname' key to the HOODAW secret
This will contain the value of `hoodaw_host`.

Once this exists, pipelines will no longer need to
hard-code the URL of the web application. Instead,
they will get both the API key and the hostname
from the same secret.

This will make it easier if we move the 
application to a different URL, because we will 
only need to change the value in one place.

Besides, hard-coded strings are icky.