build(reporting): add new infra test image and ginkgo commands
We've restructured the integration tests to ensure the tests aren't baked into a binary, thus allowing the pipeline to test everything in the main git branch. With this change comes the requirement to change how the pipeline executes. This commit uses a new docker image (created upstream) and uses the built-in ginkgo v1 binary to execute the tests against the live cluster.