fix(infrastructure): change how concourse reports to github
This change should report to github in two separate jobs rather than one. By changing where to failure occurs I'm hoping that in the event of one of the tasks failing, concourse will then send a failure to GitHub. The context_base and context naming is to allow both live and manager clusters to report their status individually.